+++ src/transformers/utils/output_capturing.py
@@ -21,6 +21,7 @@ from __future__ import annotations
 import threading
 from contextvars import ContextVar
 from dataclasses import dataclass
+from functools import wraps
 from typing import TYPE_CHECKING
 
 from .import_utils import is_torchdynamo_compiling, requires
@@ -200,3 +201,86 @@ def maybe_install_capturing_hooks(model: PreTrainedModel) -> None:
             return
         # This will install the hooks and mark the model as hooked
         install_all_output_capturing_hooks(model)
+
+
+def capture_outputs(func=None, *, tie_last_hidden_states=True):
+    """
+    Decorator to intercept specific layer outputs through hooks. The hooks are installed only once and lazily,
+    the first time output capture is requested with the `output_xxx` kwargs/config.
+    The implementation is fully context/thread safe, except when using `torch.compile`, as dynamo is unable to trace
+    through `ContextVar` methods.
+
+    Args:
+        tie_last_hidden_states (`bool`, *optional*, defaults to `True`):
+            Whether to overwrite `out.hidden_states[-1]` with the `out.last_hidden_state`.
+            This is true for all language models and should be toggled off only if
+            `out.hidden_states[-1]` has to be the hidden state before last layer norm, which
+            is needed for some vision models (e.g. CLIP, SigLIP)
+    """
+
+    def wrapped_fn(func):
+        @wraps(func)
+        def wrapper(self, *args, **kwargs):
+            # Pop it so that internal modules always return a dict even if False is requested
+            return_dict = kwargs.pop("return_dict", getattr(self.config, "return_dict", True))
+
+            # In case `kwargs` are passed as an explicit dict
+            all_passed_kargs = kwargs.copy()
+            if "kwargs" in all_passed_kargs:
+                for k, v in all_passed_kargs["kwargs"].items():
+                    all_passed_kargs[k] = v
+
+            # _can_record_outputs is None by default
+            capturable_flags = _CAN_RECORD_REGISTRY.get(str(self.__class__)) or {}
+            recordable_keys = {
+                f"output_{k}": all_passed_kargs.pop(f"output_{k}", getattr(self.config, f"output_{k}", False))
+                for k in capturable_flags
+            }
+
+            collected_outputs = {k.replace("output_", ""): [] for k, v in recordable_keys.items() if v}
+            # Make sure hooks are installed if we need to collect outputs
+            if len(collected_outputs) > 0:
+                maybe_install_capturing_hooks(self)
+            # Let's activate the output collector hooks if needed!
+            output_token = _active_collector.set(collected_outputs)
+
+            # Run the forward
+            try:
+                outputs = func(self, *args, **kwargs)
+            # Reset the states
+            finally:
+                _active_collector.reset(output_token)
+
+            # Inject collected outputs into model output (return everything as tuples for BC)
+            for key in collected_outputs:
+                if key == "hidden_states":
+                    if not tie_last_hidden_states:
+                        pass
+                    elif hasattr(outputs, "vision_hidden_states"):
+                        collected_outputs[key] = collected_outputs[key][:-1]
+                        collected_outputs[key].append(outputs.vision_hidden_states)
+                    elif hasattr(outputs, "last_hidden_state"):
+                        collected_outputs[key] = collected_outputs[key][:-1]
+                        collected_outputs[key].append(outputs.last_hidden_state)
+
+                    outputs[key] = tuple(collected_outputs[key])
+                elif key == "attentions":
+                    # In this case, the second item are cross attentions
+                    if isinstance(capturable_flags[key], list) and len(capturable_flags[key]) == 2:
+                        outputs[key] = tuple(collected_outputs[key][0::2])
+                        outputs["cross_" + key] = tuple(collected_outputs[key][1::2])
+                    else:
+                        outputs[key] = tuple(collected_outputs[key])
+                else:
+                    outputs[key] = tuple(collected_outputs[key])
+
+            if return_dict is False:
+                outputs = outputs.to_tuple()
+
+            return outputs
+
+        return wrapper
+
+    if func is not None:
+        return wrapped_fn(func)
+    return wrapped_fn
